refactor(layout): extract sidebar into reusable components
