+++ contexts/Auth/index.ts
@@ -1,3 +1,4 @@
+export type { ProtectedComponentProps } from './ProtectedComponent';
 export { RoleEnum } from '@/services/modules/users';
 export { default as ProtectedComponent } from './ProtectedComponent';
 export * from './AuthContext';
