+++ layout/components/Sidebar/SidebarItem.tsx
@@ -1,0 +1,36 @@
+import useSmoothIntoView from '@/hooks/useSmoothIntoView';
+import { cn } from '@/utils/cn';
+import { activeClass, defaultClass, disableClass } from './constant';
+
+interface SidebarItemProps extends React.HTMLAttributes<HTMLDivElement> {
+  isActive?: boolean;
+  isDisabled?: boolean;
+}
+
+function SidebarItem({
+  children,
+  className,
+  isActive,
+  isDisabled,
+  ...props
+}: SidebarItemProps) {
+  const ref = useSmoothIntoView<HTMLDivElement>();
+  return (
+    <div
+      ref={ref}
+      className={cn(
+        defaultClass,
+        isActive && activeClass,
+        isDisabled && disableClass,
+        'w-full p-0',
+        className
+      )}
+      data-active={isActive}
+      {...props}
+    >
+      {children}
+    </div>
+  );
+}
+
+export default SidebarItem;
