+++ layout/components/Sidebar/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Sidebar';
+export * from './type';
